macOS integration test fix handle /private paths properly.
On macOS the temp path is a symlink and this was making the `/snapshot`
resolution (`this.root` vs `/snapshot` in the zipFs) not work properly.

This feels a bit like a workaround, but the path being different to the
executable was the culprit in the tests not working. As that's manual
code in the integration test script, and due to the custom entrypoint
for Mocha, that makes this not a workaround :-)